board UI updates: always show search for boards and images if a term is entered, clear search when view is toggled off
