+++ invokeai/frontend/web/src/features/gallery/components/ImageGalleryContent.tsx
@@ -1,57 +1,28 @@
-import type { ChakraProps } from '@invoke-ai/ui-library';
-import {
-  Box,
-  Collapse,
-  Divider,
-  Flex,
-  IconButton,
-  Spacer,
-  Tab,
-  TabList,
-  Tabs,
-  useDisclosure,
-} from '@invoke-ai/ui-library';
+import { Box, Collapse, Divider, Flex, IconButton, useDisclosure } from '@invoke-ai/ui-library';
 import { useAppDispatch, useAppSelector } from 'app/store/storeHooks';
 import { GalleryHeader } from 'features/gallery/components/GalleryHeader';
-import { galleryViewChanged } from 'features/gallery/store/gallerySlice';
+import { boardSearchTextChanged } from 'features/gallery/store/gallerySlice';
 import ResizeHandle from 'features/ui/components/tabs/ResizeHandle';
 import { usePanel, type UsePanelOptions } from 'features/ui/hooks/usePanel';
 import type { CSSProperties } from 'react';
-import { memo, useCallback, useMemo, useRef } from 'react';
+import { memo, useCallback, useEffect, useMemo, useRef } from 'react';
 import { useTranslation } from 'react-i18next';
-import { PiMagnifyingGlassBold } from 'react-icons/pi';
+import { MdSearch, MdSearchOff } from 'react-icons/md';
 import type { ImperativePanelGroupHandle } from 'react-resizable-panels';
 import { Panel, PanelGroup } from 'react-resizable-panels';
 
 import BoardsListWrapper from './Boards/BoardsList/BoardsListWrapper';
 import BoardsSearch from './Boards/BoardsList/BoardsSearch';
+import { Gallery } from './Gallery';
 import GallerySettingsPopover from './GallerySettingsPopover/GallerySettingsPopover';
-import GalleryImageGrid from './ImageGrid/GalleryImageGrid';
-import { GalleryPagination } from './ImageGrid/GalleryPagination';
-import { GallerySearch } from './ImageGrid/GallerySearch';
 
-const COLLAPSE_STYLES: CSSProperties = { flexShrink: 0, minHeight: 0 };
-
-const BASE_STYLES: ChakraProps['sx'] = {
-  fontWeight: 'semibold',
-  fontSize: 'sm',
-  color: 'base.300',
-};
-
-const SELECTED_STYLES: ChakraProps['sx'] = {
-  borderColor: 'base.800',
-  borderBottomColor: 'base.900',
-  color: 'invokeBlue.300',
-};
+export const COLLAPSE_STYLES: CSSProperties = { flexShrink: 0, minHeight: 0 };
 
 const ImageGalleryContent = () => {
   const { t } = useTranslation();
-  const galleryView = useAppSelector((s) => s.gallery.galleryView);
-  const searchTerm = useAppSelector((s) => s.gallery.searchTerm);
   const boardSearchText = useAppSelector((s) => s.gallery.boardSearchText);
   const dispatch = useAppDispatch();
-  const searchDisclosure = useDisclosure({ defaultIsOpen: false });
-  const boardSearchDisclosure = useDisclosure({ defaultIsOpen: false });
+  const boardSearchDisclosure = useDisclosure({ defaultIsOpen: !!boardSearchText.length });
   const panelGroupRef = useRef<ImperativePanelGroupHandle>(null);
 
   const boardsListPanelOptions = useMemo<UsePanelOptions>(
@@ -67,13 +38,20 @@ const ImageGalleryContent = () => {
   );
   const boardsListPanel = usePanel(boardsListPanelOptions);
 
-  const handleClickImages = useCallback(() => {
-    dispatch(galleryViewChanged('images'));
-  }, [dispatch]);
+  const handleClickBoardSearch = useCallback(() => {
+    if (boardSearchText.length) {
+      dispatch(boardSearchTextChanged(''));
+      boardSearchDisclosure.onToggle();
+    } else {
+      boardSearchDisclosure.onToggle();
+    }
+  }, [boardSearchText, dispatch, boardSearchDisclosure]);
 
-  const handleClickAssets = useCallback(() => {
-    dispatch(galleryViewChanged('assets'));
-  }, [dispatch]);
+  useEffect(() => {
+    if (boardSearchDisclosure.isOpen) {
+      boardsListPanel.expand();
+    }
+  }, [boardSearchDisclosure, boardsListPanel]);
 
   return (
     <Flex position="relative" flexDirection="column" h="full" w="full" pt={2}>
@@ -84,25 +62,17 @@ const ImageGalleryContent = () => {
           <IconButton
             w="full"
             h="full"
-            onClick={boardSearchDisclosure.onToggle}
-            tooltip={`${t('gallery.displayBoardSearch')}`}
+            onClick={handleClickBoardSearch}
+            tooltip={
+              boardSearchDisclosure.isOpen ? `${t('gallery.exitBoardSearch')}` : `${t('gallery.displayBoardSearch')}`
+            }
             aria-label={t('gallery.displayBoardSearch')}
-            icon={<PiMagnifyingGlassBold />}
+            icon={boardSearchText.length ? <MdSearchOff /> : <MdSearch />}
             variant="link"
           />
-          {boardSearchText && (
-            <Box
-              position="absolute"
-              w={2}
-              h={2}
-              bg="invokeBlue.300"
-              borderRadius="full"
-              insetBlockStart={2}
-              insetInlineEnd={2}
-            />
-          )}
         </Box>
       </Flex>
+
       <PanelGroup ref={panelGroupRef} direction="vertical">
         <Panel
           id="boards-list-panel"
@@ -127,50 +97,7 @@ const ImageGalleryContent = () => {
           onDoubleClick={boardsListPanel.onDoubleClickHandle}
         />
         <Panel id="gallery-wrapper-panel" minSize={20}>
-          <Flex flexDirection="column" alignItems="center" justifyContent="space-between" h="full" w="full">
-            <Tabs index={galleryView === 'images' ? 0 : 1} variant="enclosed" display="flex" flexDir="column" w="full">
-              <TabList gap={2} fontSize="sm" borderColor="base.800">
-                <Tab sx={BASE_STYLES} _selected={SELECTED_STYLES} onClick={handleClickImages} data-testid="images-tab">
-                  {t('parameters.images')}
-                </Tab>
-                <Tab sx={BASE_STYLES} _selected={SELECTED_STYLES} onClick={handleClickAssets} data-testid="assets-tab">
-                  {t('gallery.assets')}
-                </Tab>
-                <Spacer />
-                <Box position="relative">
-                  <IconButton
-                    w="full"
-                    h="full"
-                    onClick={searchDisclosure.onToggle}
-                    tooltip={`${t('gallery.displaySearch')}`}
-                    aria-label={t('gallery.displaySearch')}
-                    icon={<PiMagnifyingGlassBold />}
-                    variant="link"
-                  />
-                  {searchTerm && (
-                    <Box
-                      position="absolute"
-                      w={2}
-                      h={2}
-                      bg="invokeBlue.300"
-                      borderRadius="full"
-                      insetBlockStart={2}
-                      insetInlineEnd={2}
-                    />
-                  )}
-                </Box>
-              </TabList>
-            </Tabs>
-            <Box w="full">
-              <Collapse in={searchDisclosure.isOpen} style={COLLAPSE_STYLES}>
-                <Box w="full" pt={2}>
-                  <GallerySearch />
-                </Box>
-              </Collapse>
-            </Box>
-            <GalleryImageGrid />
-            <GalleryPagination />
-          </Flex>
+          <Gallery />
         </Panel>
       </PanelGroup>
     </Flex>
